+++ bdr_registry/settings.py
@@ -1,10 +1,30 @@
 import os
 import sys
+from getenv import env
+import ldap
+from django_auth_ldap.config import LDAPSearch
+
+class secure_str(str):
+    """
+    A string that doesn't print its contents on `repr()`. Useful to
+    protect passwords.
+    """
+    __slots__ = ()
+    __repr__ = object.__repr__
+
 
 BASE_DIR = os.path.dirname(os.path.dirname(__file__))
 
-DEBUG = False
-ADMIN_ALL_BDR_TABLES = DEBUG
+DEBUG = env('DEBUG', False)
+ALLOWED_HOSTS = env('ALLOWED_HOSTS', ['localhost', '127.0.0.1'])
+ADMIN_ALL_BDR_TABLES = env('ADMIN_ALL_BDR_TABLES', False)
+
+DATABASES = env('DATABASES', {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': './db.sqlite',  # path to database file
+    }
+})
 
 ASSETS_DEBUG = True
 
@@ -28,19 +48,19 @@ MEDIA_ROOT = ''
 
 MEDIA_URL = ''
 
-STATIC_URL = '/static/'
-
-STATIC_ROOT = os.path.join(BASE_DIR, 'static')
+STATIC_ROOT = env('STATIC_ROOT', os.path.join(BASE_DIR, 'static'))
+STATIC_URL = env('STATIC_URL', '/static/')
 
 STATICFILES_FINDERS = (
     'django.contrib.staticfiles.finders.FileSystemFinder',
     'django.contrib.staticfiles.finders.AppDirectoriesFinder',
 )
 
-SITE_URL = '/'
+SITE_URL = env('SITE_URL', '/')
 LOGIN_REDIRECT_URL = SITE_URL
+SECRET_KEY = env('SECRET_KEY', 'moo4ge4F')
 
-BDR_REVERSE_PROXY = False
+ACCOUNTS_PREFIX = ''
 
 MIDDLEWARE_CLASSES = (
     'frame.middleware.RequestMiddleware',
@@ -126,17 +146,55 @@ LOGGING = {
     }
 }
 
+BDR_REVERSE_PROXY = env('BDR_REVERSE_PROXY', False)
+BDR_SERVER_URL = env('BDR_SERVER_URL', 'http://example.com/')
 
-BDR_SERVER_URL = 'https://bdr.eionet.europa.eu/'
-BDR_HELPDESK_EMAIL = ''
+# email addresses
+BDR_HELPDESK_EMAIL = env('BDR_HELPDESK_EMAIL', '')
+BDR_EMAIL_FROM = env('BDR_EMAIL_FROM', 'bdr@localhost')
 
-EMAIL_BACKEND = 'post_office.EmailBackend'
+# email server
+EMAIL_BACKEND = env('EMAIL_BACKEND', 'post_office.EmailBackend')
+EMAIL_HOST = env('EMAIL_HOST', 'localhost')
+EMAIL_PORT = env('EMAIL_PORT', 25)
 
-BDR_EMAIL_FROM = 'bdr@localhost'
-BDR_REPORTEK_ORGANISATION_URL = '#'
-BDR_API_URL = None
-BDR_API_AUTH = None
-BDR_AUDIT_LOG_FILE = None
+# authentication using LDAP
+AUTH_LDAP_SERVER_URI = env('AUTH_LDAP_SERVER_URI', '')
+AUTH_LDAP_BIND_DN = env('AUTH_LDAP_BIND_DN', '')
+AUTH_LDAP_BIND_PASSWORD = secure_str(env('AUTH_LDAP_BIND_PASSWORD', ''))
+AUTH_LDAP_BASE_DN = env('AUTH_LDAP_BASE_DN', '')
+AUTH_LDAP_USER_SEARCH = LDAPSearch(AUTH_LDAP_BASE_DN,
+                                  ldap.SCOPE_SUBTREE,
+                                  '(uid=%(user)s)')
+
+AUTH_LDAP_USER_ATTR_MAP = env('AUTH_LDAP_USER_ATTR_MAP', {
+   "first_name": "givenName",
+   "last_name": "sn",
+   "email": "mail",
+})
+
+AUTHENTICATION_BACKENDS = env('AUTHENTICATION_BACKENDS', (
+   'django_auth_ldap.backend.LDAPBackend',
+   'django.contrib.auth.backends.ModelBackend',
+   'frame.backends.FrameUserBackend',
+))
+
+FRAME_URL = env('FRAME_URL', '')
+FRAME_VERIFY_SSL = env('FRAME_VERIFY_SSL', False)
+FRAME_COOKIES = env('FRAME_COOKIES', [])
+
+# credentials for updating LDAP accounts
+LDAP_EDIT_SERVER = AUTH_LDAP_SERVER_URI
+LDAP_EDIT_DN = env('LDAP_EDIT_DN', '')
+LDAP_EDIT_PASSWORD = secure_str(env('LDAP_EDIT_PASSWORD', ''))
+
+BDR_REPORTEK_ORGANISATION_URL = env('BDR_REPORTEK_ORGANISATION_URL', '#')
+BDR_API_URL = env('BDR_API_PORTAL_URL', 'http://example.com/api')
+BDR_AUDIT_LOG_FILE = env('BDR_AUDIT_LOG_FILE', None)
+
+BDR_API_AUTH_USER = env('BDR_API_AUTH_USER', '')
+BDR_API_AUTH_PASSWORD = env('BDR_API_AUTH_PASSWORD', '')
+BDR_API_AUTH = (BDR_API_AUTH_USER, secure_str(BDR_API_AUTH_PASSWORD))
 
 DATE_FORMAT = '%d %b %Y'