+++ bdr_registry/test_settings.py
@@ -20,3 +20,6 @@ DATABASES = {
 EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
 EMAIL_HOST = 'localhost'
 EMAIL_PORT = 25
+
+BDR_API_URL = 'http://example.com/api'
+REPORTING_YEAR = 2016